env_requirements.txt
bcbio-gff==0.6.6
biopython==1.78
numpy==1.19.2
six==1.15.0
